==> icare_code/1.1_snomed_codes.ipynb <==
# Libraries
import pandas as pd
pd.set_option('display.max_columns', 40)
pd.set_option('display.width', 2000)


from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

%pip install \
--trusted-host 192.168.210.244 \
--index-url http://192.168.210.244:8081/repository/pypi/simple/ \
snowflake-connector-python

import json, snowflake.connector

# establish the connection to snowflake
ctx = snowflake.connector.connect( 
    **json.load(open('/opt/ich/python-snowflake-defaults.json')))
    
# verify and test if connection is working
try: 
    cs = ctx.cursor() 
    cs.execute('SELECT current_version(), current_role(), current_warehouse()')
    print(cs.fetchone())
finally: 
    cs.close()

## Problem

# Import 
query = '''
SELECT * from ICHT_PROD.ICHT_COVID.PROBLEM
'''
cur = ctx.cursor().execute(query)
problem_df = pd.DataFrame.from_records(iter(cur), columns=[x[0] for x in cur.description])

# Filter for disorders
filtered_problem_df = problem_df[['PROBLEM', 'PROBLEM_DESC']]
filtered_problem_df = filtered_problem_df[filtered_problem_df['PROBLEM_DESC'].str.contains("\(disorder\)")==True]
filtered_problem_df.drop_duplicates(inplace=True)
filtered_problem_df.reset_index(inplace=True, drop=True)

# Get count
counts = problem_df[['PROBLEM']].value_counts()
counts = counts.reset_index()
counts.columns = ['PROBLEM', 'count']
# Merge
filtered_problem_df2 = pd.merge(filtered_problem_df, counts, how='left', on='PROBLEM')

# Drop duplicate
filtered_problem_df2.drop_duplicates(subset=['PROBLEM'], inplace=True)
# Sort 
filtered_problem_df2.sort_values(by=['count'], ascending=False, inplace=True)
filtered_problem_df2.reset_index(inplace=True, drop=True)

# Drop none
filtered_problem_df3 = filtered_problem_df2[filtered_problem_df2['PROBLEM'].str.contains("ICD")==False]

# Filter for move than 5
filtered_problem_df4 = filtered_problem_df3[filtered_problem_df3['count'] >= 5]

# Save
filtered_problem_df4.to_csv('snomed_problem_codes_morethan5.csv', index=False)

## Riskfactor

# Import 
query = '''
SELECT * from ICHT_PROD.ICHT_COVID.RISKFACTORS
'''
cur = ctx.cursor().execute(query)
riskfactor_df = pd.DataFrame.from_records(iter(cur), columns=[x[0] for x in cur.description])

# Filter for disorders
filtered_riskfactor_df = riskfactor_df[['SNOMED_CODE', 'CODE_DESCRIPTION']]
filtered_riskfactor_df = filtered_riskfactor_df[filtered_riskfactor_df['CODE_DESCRIPTION'].str.contains("\(disorder\)")==True]
filtered_riskfactor_df.drop_duplicates(inplace=True)
filtered_riskfactor_df.reset_index(inplace=True, drop=True)

# Get count
counts = riskfactor_df[['SNOMED_CODE']].value_counts()
counts = counts.reset_index()
counts.columns = ['SNOMED_CODE', 'count']
# Merge
filtered_riskfactor_df2 = pd.merge(filtered_riskfactor_df, counts, how='left', on='SNOMED_CODE')

# Drop duplicates
filtered_riskfactor_df2.drop_duplicates(subset=['SNOMED_CODE'], inplace=True)
# Sort 
filtered_riskfactor_df2.sort_values(by=['count'], ascending=False, inplace=True)
filtered_riskfactor_df2.reset_index(inplace=True, drop=True)

# Drop none
filtered_riskfactor_df3 = filtered_riskfactor_df2[filtered_riskfactor_df2['SNOMED_CODE'].str.contains("ICD")==False]

# Filter for move than 5
filtered_riskfactor_df4 = filtered_riskfactor_df3[filtered_riskfactor_df3['count'] >= 5]

# Save
filtered_riskfactor_df4.to_csv('snomed_riskfactor_codes_morethan5.csv', index=False)

## Diagnosis

# Import 
query = '''
SELECT * from ICHT_PROD.ICHT_COVID.EPISODES_DIAGNOSIS
'''
cur = ctx.cursor().execute(query)
diagnosis_df = pd.DataFrame.from_records(iter(cur), columns=[x[0] for x in cur.description])

# Filter for disorders
filtered_diagnosis_df = diagnosis_df[['DIAGNOSIS_CODE_SNOMED', 'DIAGNOSIS_DESCRIPTION_SNOMED']]
filtered_diagnosis_df = filtered_diagnosis_df[filtered_diagnosis_df['DIAGNOSIS_DESCRIPTION_SNOMED'].str.contains("\(disorder\)")==True]
filtered_diagnosis_df.drop_duplicates(inplace=True)
filtered_diagnosis_df.reset_index(inplace=True, drop=True)
# Get count
counts = diagnosis_df[['DIAGNOSIS_CODE_SNOMED']].value_counts()
counts = counts.reset_index()
counts.columns = ['DIAGNOSIS_CODE_SNOMED', 'count']
# Merge
filtered_diagnosis_df2 = pd.merge(filtered_diagnosis_df, counts, how='left', on='DIAGNOSIS_CODE_SNOMED')
# Drop duplicates
filtered_diagnosis_df2.drop_duplicates(subset=['DIAGNOSIS_CODE_SNOMED'], inplace=True)
# Sort 
filtered_diagnosis_df2.sort_values(by=['count'], ascending=False, inplace=True)
filtered_diagnosis_df2.reset_index(inplace=True, drop=True)
# Drop none
filtered_diagnosis_df3 = filtered_diagnosis_df2[filtered_diagnosis_df2['DIAGNOSIS_CODE_SNOMED'].str.contains("ICD")==False]
# Filter for move than 5
filtered_diagnosis_df4 = filtered_diagnosis_df3[filtered_diagnosis_df3['count'] >= 5]

# Remove SNOMED string
filtered_diagnosis_df4['DIAGNOSIS_CODE_SNOMED'] = filtered_diagnosis_df4['DIAGNOSIS_CODE_SNOMED'].str.replace('SNOMED!', '')

# Save
filtered_diagnosis_df4.to_csv('snomed_diagnosis_codes_morethan5.csv', index=False)
