==> icare_code/2.2_med_infilling.ipynb <==
# Libraries
import pandas as pd
pd.set_option('display.max_columns', 40)
pd.set_option('display.width', 2000)

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

%pip install \
--trusted-host 192.168.210.244 \
--index-url http://192.168.210.244:8081/repository/pypi/simple/ \
snowflake-connector-python

import json, snowflake.connector

# establish the connection to snowflake
ctx = snowflake.connector.connect( 
    **json.load(open('/opt/ich/python-snowflake-defaults.json')))
    
# verify and test if connection is working
try: 
    cs = ctx.cursor() 
    cs.execute('SELECT current_version(), current_role(), current_warehouse()')
    print(cs.fetchone())
finally: 
    cs.close()

# Import
path = r'data/drugs_indications_and_codes.csv'
drugs_indications_and_codes = pd.read_csv(path)

path = r'data/final_problem_dummies.csv'
final_problem_dummies = pd.read_csv(path)

# Import 
query = '''
SELECT * from ICHT_PROD.ICHT_COVID.PHARMACY_PRESCRIBING
'''
cur = ctx.cursor().execute(query)
prescribing_df = pd.DataFrame.from_records(iter(cur), columns=[x[0] for x in cur.description])

# Filter for administration
prescribing_df = prescribing_df[prescribing_df['ADMISSION_MEDICINE_Y_N'] == 'Yes']
prescribing_df = prescribing_df[['SUBJECT', 'PRESCRIPTION_ORDER_ID','ORDER_DT_TM', 'PRESCRIPTION_TYPE', 'THERAPEUTICAL_CLASS', 'MEDICATION_NAME_SHORT', 'MEDICATION_NAME']]

# Filter for those subjects of interest
prescribing_df = prescribing_df[prescribing_df['SUBJECT'].isin(final_problem_dummies['SUBJECT'].to_list())]

# Filter for medications of interest
prescribing_df = prescribing_df[prescribing_df['MEDICATION_NAME_SHORT'].isin(drugs_indications_and_codes['icht_prescribing_medication'].to_list())]

# Save
#prescribing_df.to_csv('filtered_pharmacy_prescribing.csv')

# Convert to datetime
final_problem_dummies['PROBLEM_DT_TM'] = pd.to_datetime(final_problem_dummies['PROBLEM_DT_TM'])
prescribing_df['ORDER_DT_TM'] = pd.to_datetime(prescribing_df['ORDER_DT_TM'])

for x in range(len(final_problem_dummies)):
    if x % 1000 == 0:
        print('x:', x)
    # Filter by subject
    temp_df = prescribing_df[prescribing_df['SUBJECT'] == final_problem_dummies.loc[x, 'SUBJECT']]
    if len(temp_df) >= 1:
        # Filter by date
        temp_df = temp_df[temp_df['ORDER_DT_TM'] <= final_problem_dummies.loc[x, 'PROBLEM_DT_TM']]
        if len(temp_df) >= 1:
            temp_drug_df = drugs_indications_and_codes[drugs_indications_and_codes['icht_prescribing_medication'].isin(temp_df['MEDICATION_NAME_SHORT'].to_list())]
            snomed_code_list = temp_drug_df['snomed_code'].astype(str).unique().tolist()
            # Add prefix 
            prefix_ADD = 'PROBLEM_'
            snomed_code_list = [prefix_ADD + z for z in snomed_code_list if isinstance(z, str)]
            # Find snomed codes that overlap
            snomed_code_list = list(set(snomed_code_list) & set(final_problem_dummies.iloc[:,3:].columns.tolist()))
            # Set those problems to 1 in main df
            final_problem_dummies.loc[x, snomed_code_list] = 1

# Save
#final_problem_dummies.to_csv('final_problem_dummies_med.csv')

## Run with fuzzy

# Import
path = r'data/drugs_indications_and_codes_fuzzy.csv'
drugs_indications_and_codes_fuzzy = pd.read_csv(path)

# Import
path = r'data/final_problem_dummies.csv'
final_problem_dummies = pd.read_csv(path)

# Convert to datetime
final_problem_dummies['PROBLEM_DT_TM'] = pd.to_datetime(final_problem_dummies['PROBLEM_DT_TM'])
prescribing_df['ORDER_DT_TM'] = pd.to_datetime(prescribing_df['ORDER_DT_TM'])

for x in range(len(final_problem_dummies)):
    if x % 1000 == 0:
        print('x:', x)
    # Filter by subject
    temp_df = prescribing_df[prescribing_df['SUBJECT'] == final_problem_dummies.loc[x, 'SUBJECT']]
    if len(temp_df) >= 1:
        # Filter by date
        temp_df = temp_df[temp_df['ORDER_DT_TM'] <= final_problem_dummies.loc[x, 'PROBLEM_DT_TM']]
        if len(temp_df) >= 1:
            temp_drug_df = drugs_indications_and_codes_fuzzy[drugs_indications_and_codes_fuzzy['icht_prescribing_medication'].isin(temp_df['MEDICATION_NAME_SHORT'].to_list())]
            snomed_code_list = temp_drug_df['snomed_code'].astype(str).unique().tolist()
            # Add prefix 
            prefix_ADD = 'PROBLEM_'
            snomed_code_list = [prefix_ADD + z for z in snomed_code_list if isinstance(z, str)]
            # Find snomed codes that overlap
            snomed_code_list = list(set(snomed_code_list) & set(final_problem_dummies.iloc[:,3:].columns.tolist()))
            # Set those problems to 1 in main df
            final_problem_dummies.loc[x, snomed_code_list] = 1

# Save
#final_problem_dummies.to_csv('final_problem_dummies_med_fuzzy.csv')
