==> icare_code/3.1_charlson_score.ipynb <==
# Libraries
import pandas as pd
pd.set_option('display.max_columns', 40)
pd.set_option('display.width', 2000)

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

# Import
path = r'data/filtered_snomed_charlson_index.csv'
filtered_snomed_charlson_index = pd.read_csv(path, dtype={'code': 'str'})

Normal problem dummies

# Import
path = r'data/final_problem_dummies.csv'
final_problem_dummies = pd.read_csv(path)

filtered_snomed_charlson_index

# Get lists
code_list = filtered_snomed_charlson_index['code'].unique().tolist()
other_list = ['new_subject', 'SUBJECT', 'DT_TM']
code_list = other_list + code_list

# Remove '420270002', '51292008' as not in axis 
code_list.remove('420270002')
code_list.remove('51292008')

# Remove problem prefix from column names
final_problem_dummies.columns = final_problem_dummies.columns.str.removeprefix('PROBLEM_')

# Filter
final_problem_dummies = final_problem_dummies[code_list]

# Create score for each patient! 
final_problem_dummies_2 = final_problem_dummies.copy()
for column in final_problem_dummies_2.iloc[:,3:].columns:
    # Replace values with weights
    final_problem_dummies_2.loc[final_problem_dummies_2[column] == 1, column] = filtered_snomed_charlson_index[filtered_snomed_charlson_index['code'] == column]['weight'].iloc[0]
final_problem_dummies_2['charlson_score'] = final_problem_dummies_2.iloc[:,3:].sum(axis=1)
final_problem_dummies_2 = final_problem_dummies_2[['new_subject', 'SUBJECT', 'DT_TM', 'charlson_score']]

final_problem_dummies_2.charlson_score.max()
final_problem_dummies_2.charlson_score.min()
final_problem_dummies_2.charlson_score.mean()
final_problem_dummies_2.charlson_score.mode()

# save
#final_problem_dummies_2.to_csv('problem_charlson_score.csv', index=False)

Problem dummies med

# Import
path = r'data/final_problem_dummies_med.csv'
final_problem_dummies_med = pd.read_csv(path)

# Get lists
code_list = filtered_snomed_charlson_index['code'].unique().tolist()
other_list = ['new_subject', 'SUBJECT', 'DT_TM']
code_list = other_list + code_list

# Remove '420270002', '51292008' as not in axis 
code_list.remove('420270002')
code_list.remove('51292008')

# Remove problem prefix from column names
final_problem_dummies_med.columns = final_problem_dummies_med.columns.str.removeprefix('PROBLEM_')

# Filter
final_problem_dummies_med = final_problem_dummies_med[code_list]

# Create score for each patient!
final_problem_dummies_med_2 = final_problem_dummies_med.copy()
for column in final_problem_dummies_med_2.iloc[:,3:].columns:
    # Replace values with weights
    final_problem_dummies_med_2.loc[final_problem_dummies_med_2[column] == 1, column] = filtered_snomed_charlson_index[filtered_snomed_charlson_index['code'] == column]['weight'].iloc[0]
final_problem_dummies_med_2['charlson_score'] = final_problem_dummies_med_2.iloc[:,3:].sum(axis=1)
final_problem_dummies_med_2 = final_problem_dummies_med_2[['new_subject', 'SUBJECT', 'DT_TM', 'charlson_score']]

# save
#final_problem_dummies_med_2.to_csv('problem_charlson_score_med.csv', index=False)

Problem dummies med fuzzy

# Import
path = r'data/final_problem_dummies_med_fuzzy.csv'
final_problem_dummies_med_fuzzy = pd.read_csv(path)

# Get lists
code_list = filtered_snomed_charlson_index['code'].unique().tolist()
other_list = ['new_subject', 'SUBJECT', 'DT_TM']
code_list = other_list + code_list
# Remove '420270002', '51292008' as not in axis 
code_list.remove('420270002')
code_list.remove('51292008')
# Remove problem prefix from column names
final_problem_dummies_med_fuzzy.columns = final_problem_dummies_med_fuzzy.columns.str.removeprefix('PROBLEM_')
# Filter
final_problem_dummies_med_fuzzy = final_problem_dummies_med_fuzzy[code_list]
# Create score for each patient!
final_problem_dummies_med_fuzzy_2 = final_problem_dummies_med_fuzzy.copy()
for column in final_problem_dummies_med_fuzzy_2.iloc[:,3:].columns:
    # Replace values with weights
    final_problem_dummies_med_fuzzy_2.loc[final_problem_dummies_med_fuzzy_2[column] == 1, column] = filtered_snomed_charlson_index[filtered_snomed_charlson_index['code'] == column]['weight'].iloc[0]
final_problem_dummies_med_fuzzy_2['charlson_score'] = final_problem_dummies_med_fuzzy_2.iloc[:,3:].sum(axis=1)
final_problem_dummies_med_fuzzy_2 = final_problem_dummies_med_fuzzy_2[['new_subject', 'SUBJECT', 'DT_TM', 'charlson_score']]

# save
#final_problem_dummies_med_fuzzy_2.to_csv('problem_charlson_score_med_fuzzy.csv', index=False)
