==> icare_code/3.2_death_label.ipynb <==
# Libraries
import pandas as pd
pd.set_option('display.max_columns', 40)
pd.set_option('display.width', 2000)

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

%pip install \
--trusted-host 192.168.210.244 \
--index-url http://192.168.210.244:8081/repository/pypi/simple/ \
snowflake-connector-python

import json, snowflake.connector

# establish the connection to snowflake
ctx = snowflake.connector.connect( 
    **json.load(open('/opt/ich/python-snowflake-defaults.json')))
    
# verify and test if connection is working
try: 
    cs = ctx.cursor() 
    cs.execute('SELECT current_version(), current_role(), current_warehouse()')
    print(cs.fetchone())
finally: 
    cs.close()

query = '''
SELECT * from ICHT_PROD.ICHT_COVID.DEMOGRAPHIC
'''

cur = ctx.cursor().execute(query)
demographic_df = pd.DataFrame.from_records(iter(cur), columns=[x[0] for x in cur.description])

Problem dummies

# Import
path = r'data/problem_charlson_score.csv'
problem_charlson_score = pd.read_csv(path)

# Merge
problem_charlson = pd.merge(problem_charlson_score, demographic_df[['SUBJECT', 'DEATH_DATE']], how='left')

# To datetime
problem_charlson['DT_TM'] = pd.to_datetime(problem_charlson['DT_TM']).dt.date
problem_charlson['DEATH_DATE'] = pd.to_datetime(problem_charlson['DEATH_DATE']).dt.date

# Work out if death within a year
def year_fun(row):
    if pd.notnull(row['DEATH_DATE']):
        if row['DEATH_DATE'] <= row['DT_TM'] + pd.offsets.DateOffset(years=1):
            return 1

# Work out if death at anytime
def death_fun(row):
    if pd.notnull(row['DEATH_DATE']):
        return 1

# Apply
problem_charlson['death_year_label'] = problem_charlson.apply (lambda row: year_fun (row),axis=1)    
problem_charlson['death_label'] = problem_charlson.apply (lambda row: death_fun (row),axis=1) 

# save
#problem_charlson.to_csv('problem_charlson_death.csv', index=False)

MED

# Import
path = r'data/problem_charlson_score_med.csv'
problem_charlson_score_med = pd.read_csv(path)

problem_charlson_med = pd.merge(problem_charlson_score_med, demographic_df[['SUBJECT', 'DEATH_DATE']], how='left')
problem_charlson_med['DT_TM'] = pd.to_datetime(problem_charlson_med['DT_TM']).dt.date
problem_charlson_med['DEATH_DATE'] = pd.to_datetime(problem_charlson_med['DEATH_DATE']).dt.date
problem_charlson_med['death_year_label'] = problem_charlson_med.apply (lambda row: year_fun (row),axis=1)    
problem_charlson_med['death_label'] = problem_charlson_med.apply (lambda row: death_fun (row),axis=1) 

# Save
#problem_charlson_med.to_csv('problem_charlson_death_med.csv', index=False)

Med fuzzy

# Import
path = r'data/problem_charlson_score_med_fuzzy.csv'
problem_charlson_score_med_fuzzy = pd.read_csv(path)

problem_charlson_med_fuzzy = pd.merge(problem_charlson_score_med_fuzzy, demographic_df[['SUBJECT', 'DEATH_DATE']], how='left')
problem_charlson_med_fuzzy['DT_TM'] = pd.to_datetime(problem_charlson_med_fuzzy['DT_TM']).dt.date
problem_charlson_med_fuzzy['DEATH_DATE'] = pd.to_datetime(problem_charlson_med_fuzzy['DEATH_DATE']).dt.date
problem_charlson_med_fuzzy['death_year_label'] = problem_charlson_med_fuzzy.apply (lambda row: year_fun (row),axis=1)    
problem_charlson_med_fuzzy['death_label'] = problem_charlson_med_fuzzy.apply (lambda row: death_fun (row),axis=1) 

# Save
#problem_charlson_med_fuzzy.to_csv('problem_charlson_death_med_fuzzy.csv', index=False)
