==> icare_code/3.3_los_labels.ipynb <==
# Libraries
import numpy as np
import pandas as pd
pd.set_option('display.max_columns', 40)
pd.set_option('display.width', 2000)

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

import json, snowflake.connector

# establish the connection to snowflake
ctx = snowflake.connector.connect( 
    **json.load(open('/opt/ich/python-snowflake-defaults.json')))
    
# verify and test if connection is working
try: 
    cs = ctx.cursor() 
    cs.execute('SELECT current_version(), current_role(), current_warehouse()')
    print(cs.fetchone())
finally: 
    cs.close()

# Import 
query = '''
SELECT * from ICHT_PROD.ICHT_COVID.EPISODES
'''
cur = ctx.cursor().execute(query)
episodes = pd.DataFrame.from_records(iter(cur), columns=[x[0] for x in cur.description])

# Filter for relevant columns
episodes = episodes[['SUBJECT', 'SPELL_IDENTIFIER',  'ADMISSION_DATE_TIME', 'DISCHARGE_DATE_TIME']]

# Import
path = r'data/final_problem_dummies.csv'
problem_dummies = pd.read_csv(path)

# Filter for relevant columns
problem_dummies2 = problem_dummies.iloc[:,:3]

# Filter for where DISCHARGE_DATE_TIME is after ADMISSION_DATE_TIME
episodes = episodes.loc[(episodes['DISCHARGE_DATE_TIME'] >= episodes['ADMISSION_DATE_TIME'] )]
# To date time 
problem_dummies2['PROBLEM_DT_TM'] = pd.to_datetime(problem_dummies2['PROBLEM_DT_TM'])
# Merge on subject
los = pd.merge(problem_dummies2, episodes)

# Filter for where ADMISSION_DATE_TIME is after PROBLEM_DT_TM
los = los.loc[(los['ADMISSION_DATE_TIME'] >= los['PROBLEM_DT_TM'])]
# Drop duplicates
los = los.drop_duplicates()

# Calculate los
los['los_delta'] = los['DISCHARGE_DATE_TIME'] - los['ADMISSION_DATE_TIME']
los['los_days'] = (los['los_delta'] / np.timedelta64(1, 'D')).astype(int)
los['los_hours'] = (los['los_delta'] / np.timedelta64(1, 'h')).astype(int)
def f(row):
    if row['los_days'] > 7:
        val = 1
    else:
        val = 0
    return val
los['los_long'] = los.apply(f, axis=1)

# Save
#los.to_csv('los.csv')
