==> icare_code/4.1_patient_embeddings.ipynb <==
# Libraries
import pandas as pd
pd.set_option('display.max_columns', 40)
pd.set_option('display.width', 2000)

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

# Import
path = r'data/final_trimmed_snomed_embedding_128d.csv'
embedding = pd.read_csv(path)

# Import
path = r'data/final_problem_dummies.csv'
problem_dummies = pd.read_csv(path)

# Conver to string
embedding['snomed_code'] = embedding['snomed_code'].astype(str)
# Drop columns
problem_dummies.drop(columns=['SUBJECT', 'PROBLEM_DT_TM'], inplace=True)
# Remove prefix
problem_dummies.columns = problem_dummies.columns.str.strip('PROBLEM_')
# Set index
problem_dummies.set_index('new_subject', inplace=True)

# Work out patient embeddings - mean
patients_embeddings = pd.DataFrame()
for index, row in problem_dummies.iterrows():
    if index % 10000 == 0:
        print(index)
    patient_df = pd.DataFrame()
    # Get problems the patient has
    codes = row[row == 1].index.tolist()
    for problem in codes:
        # Get embedding of the problem 
        temp_df = embedding[embedding['snomed_code'] == problem]
        patient_df = pd.concat([patient_df, temp_df.drop(columns=['snomed_code'])], ignore_index=True)
       
    temp_df = pd.DataFrame(patient_df.mean()).transpose()
    patients_embeddings = pd.concat([patients_embeddings, temp_df], ignore_index=True)
patients_embeddings.index = problem_dummies.index

# Save
#patients_embeddings.to_csv('trimmed_patient_embedding_128d.csv')

# Work out patient embeddings - sum
patients_embeddings = pd.DataFrame()
for index, row in problem_dummies.iterrows():
    if index % 10000 == 0:
        print(index)
    patient_df = pd.DataFrame()
    # Get problems the patient has
    codes = row[row == 1].index.tolist()
    for problem in codes:
        # Get embedding of the problem 
        temp_df = embedding[embedding['snomed_code'] == problem]
        patient_df = pd.concat([patient_df, temp_df.drop(columns=['snomed_code'])], ignore_index=True)
       
    temp_df = pd.DataFrame(patient_df.sum()).transpose()
    patients_embeddings = pd.concat([patients_embeddings, temp_df], ignore_index=True)
patients_embeddings.index = problem_dummies.index

# Save
#patients_embeddings.to_csv('trimmed_patient_embedding_128d_sum.csv')
