==> icare_code/6.2_cbr_charlson_metric.ipynb <==
# Libraries
import numpy as np
import pandas as pd
pd.set_option('display.max_columns', 40)
pd.set_option('display.width', 2000)

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

# Import
path = r'data/filtered_snomed_charlson_index.csv'
filtered_snomed_charlson_index = pd.read_csv(path, dtype={'code': 'str'})

filtered_snomed_charlson_index

filtered_snomed_charlson_index.comorbid_condition.value_counts()
filtered_snomed_charlson_index.comorbid_condition.nunique()
filtered_snomed_charlson_index.code.nunique()

# Import
path = r'data/cbr_metric_my_method.csv'
my_method = pd.read_csv(path)

path = r'data/cbr_metric_one_hot.csv'
one_hot = pd.read_csv(path)

path = r'data/cbr_metric_rocheteau.csv'
rocheteau = pd.read_csv(path)

# Import
path = r'data/final_problem_dummies.csv'
problem_dummies = pd.read_csv(path)
# Drop columns
problem_dummies.drop(columns=['SUBJECT', 'PROBLEM_DT_TM'], inplace=True)
# Remove prefix
problem_dummies.columns = problem_dummies.columns.str.strip('PROBLEM_')
# Set index
problem_dummies.set_index('new_subject', inplace=True)

# Get charlson_jaccard_index for my_method
my_method_score_list = []

for index, row in my_method.iterrows():
    patient_n = row['0']
    neighbour_n = row['1']
    # Get patient and neighbour problems
    patient_problems = problem_dummies.loc[patient_n][problem_dummies.loc[patient_n] == 1].index.tolist()
    neighbour_problems = problem_dummies.loc[neighbour_n][problem_dummies.loc[neighbour_n] == 1].index.tolist()
    # Get charlson catergories
    patient_charlson = filtered_snomed_charlson_index[filtered_snomed_charlson_index['code'].isin(patient_problems)]
    neighbour_charlson = filtered_snomed_charlson_index[filtered_snomed_charlson_index['code'].isin(neighbour_problems)]
    if patient_charlson.empty:
        patient_charlson = []
    else:
        patient_charlson = patient_charlson['comorbid_condition'].unique().tolist()
    if neighbour_charlson.empty:
        neighbour_charlson = []
    else:
        neighbour_charlson = neighbour_charlson['comorbid_condition'].unique().tolist()
    # If both are empty return 1
    if not patient_charlson and not neighbour_charlson:
        my_method_score_list.append(1)
    else:
        # Work out jaccard similarity
        s1 = set(patient_charlson)
        s2 = set(neighbour_charlson)
        my_method_score_list.append(float(len(s1.intersection(s2)) / len(s1.union(s2))))

len(my_method_score_list)
np.nanmean(my_method_score_list)
np.nanstd(my_method_score_list)

# Get jaccard_index for my_method
my_method_jaccard_score_list = []

for index, row in my_method.iterrows():
    patient_n = row['0']
    neighbour_n = row['1']
    # Get patient and neighbour problems
    patient_problems = problem_dummies.loc[patient_n][problem_dummies.loc[patient_n] == 1].index.tolist()
    neighbour_problems = problem_dummies.loc[neighbour_n][problem_dummies.loc[neighbour_n] == 1].index.tolist()
    # Work out jaccard similarity
    s1 = set(patient_problems)
    s2 = set(neighbour_problems)
    my_method_jaccard_score_list.append(float(len(s1.intersection(s2)) / len(s1.union(s2))))

len(my_method_jaccard_score_list)
np.nanmean(my_method_jaccard_score_list)
np.nanstd(my_method_jaccard_score_list)

# Get charlson_jaccard_index for one_hot
one_hot_score_list = []

for index, row in one_hot.iterrows():
    patient_n = row['0']
    neighbour_n = row['1']
    # Get patient and neighbour problems
    patient_problems = problem_dummies.loc[patient_n][problem_dummies.loc[patient_n] == 1].index.tolist()
    neighbour_problems = problem_dummies.loc[neighbour_n][problem_dummies.loc[neighbour_n] == 1].index.tolist()
    # Get charlson catergories
    patient_charlson = filtered_snomed_charlson_index[filtered_snomed_charlson_index['code'].isin(patient_problems)]
    neighbour_charlson = filtered_snomed_charlson_index[filtered_snomed_charlson_index['code'].isin(neighbour_problems)]
    if patient_charlson.empty:
        patient_charlson = []
    else:
        patient_charlson = patient_charlson['comorbid_condition'].unique().tolist()
    if neighbour_charlson.empty:
        neighbour_charlson = []
    else:
        neighbour_charlson = neighbour_charlson['comorbid_condition'].unique().tolist()
    # If both are empty return 1
    if not patient_charlson and not neighbour_charlson:
        one_hot_score_list.append(1)
    else:
        # Work out jaccard similarity
        s1 = set(patient_charlson)
        s2 = set(neighbour_charlson)
        one_hot_score_list.append(float(len(s1.intersection(s2)) / len(s1.union(s2))))

len(one_hot_score_list)
np.nanmean(one_hot_score_list)
np.nanstd(one_hot_score_list)

# Get jaccard_index for one_hot
one_hot_jaccard_score_list = []

for index, row in one_hot.iterrows():
    patient_n = row['0']
    neighbour_n = row['1']
    # Get patient and neighbour problems
    patient_problems = problem_dummies.loc[patient_n][problem_dummies.loc[patient_n] == 1].index.tolist()
    neighbour_problems = problem_dummies.loc[neighbour_n][problem_dummies.loc[neighbour_n] == 1].index.tolist()
    # Work out jaccard similarity
    s1 = set(patient_problems)
    s2 = set(neighbour_problems)
    one_hot_jaccard_score_list.append(float(len(s1.intersection(s2)) / len(s1.union(s2))))

len(one_hot_jaccard_score_list)
np.nanmean(one_hot_jaccard_score_list)
np.nanstd(one_hot_jaccard_score_list)

# Get charlson_jaccard_index for rocheteau
rocheteau_score_list = []

for index, row in rocheteau.iterrows():
    patient_n = row['0']
    neighbour_n = row['1']
    # Get patient and neighbour problems
    patient_problems = problem_dummies.loc[patient_n][problem_dummies.loc[patient_n] == 1].index.tolist()
    neighbour_problems = problem_dummies.loc[neighbour_n][problem_dummies.loc[neighbour_n] == 1].index.tolist()
    # Get charlson catergories
    patient_charlson = filtered_snomed_charlson_index[filtered_snomed_charlson_index['code'].isin(patient_problems)]
    neighbour_charlson = filtered_snomed_charlson_index[filtered_snomed_charlson_index['code'].isin(neighbour_problems)]
    if patient_charlson.empty:
        patient_charlson = []
    else:
        patient_charlson = patient_charlson['comorbid_condition'].unique().tolist()
    if neighbour_charlson.empty:
        neighbour_charlson = []
    else:
        neighbour_charlson = neighbour_charlson['comorbid_condition'].unique().tolist()
    # If both are empty return 1
    if not patient_charlson and not neighbour_charlson:
        rocheteau_score_list.append(1)
    else:
        # Work out jaccard similarity
        s1 = set(patient_charlson)
        s2 = set(neighbour_charlson)
        rocheteau_score_list.append(float(len(s1.intersection(s2)) / len(s1.union(s2))))

len(rocheteau_score_list)
np.nanmean(rocheteau_score_list)
np.nanstd(rocheteau_score_list)

# Get jaccard_index for rocheteau
rocheteau_jaccard_score_list = []

for index, row in rocheteau.iterrows():
    patient_n = row['0']
    neighbour_n = row['1']
    # Get patient and neighbour problems
    patient_problems = problem_dummies.loc[patient_n][problem_dummies.loc[patient_n] == 1].index.tolist()
    neighbour_problems = problem_dummies.loc[neighbour_n][problem_dummies.loc[neighbour_n] == 1].index.tolist()
    # Work out jaccard similarity
    s1 = set(patient_problems)
    s2 = set(neighbour_problems)
    rocheteau_jaccard_score_list.append(float(len(s1.intersection(s2)) / len(s1.union(s2))))

len(rocheteau_jaccard_score_list)
np.nanmean(rocheteau_jaccard_score_list)
np.nanstd(rocheteau_jaccard_score_list)

Test stats

from scipy.stats import shapiro 
from scipy.stats import kstest
from scipy.stats import ttest_rel
from scipy.stats import wilcoxon

# Test if normally distributed (wont be)
shapiro(my_method_score_list)
kstest(my_method_score_list, 'norm')
shapiro(one_hot_score_list)
kstest(one_hot_score_list, 'norm')
shapiro(rocheteau_score_list)
kstest(rocheteau_score_list, 'norm')

import scipy.stats as stats

def test_stats(list1, list2):

    k2, p = stats.wilcoxon(list1, list2)
    alpha = 0.05
    print(p)
    if p < alpha:
        print('Different distribution')
    else:
        print(' Same distribution')

test_stats(my_method_score_list, rocheteau_score_list)

test_stats(my_method_score_list, one_hot_score_list)

test_stats(rocheteau_score_list, one_hot_score_list)

Save

# Add columns
my_method['charlson_jaccard_score'] = my_method_score_list
my_method['jaccard_score'] = my_method_jaccard_score_list
one_hot['charlson_jaccard_score'] = one_hot_score_list
one_hot['jaccard_score'] = one_hot_jaccard_score_list
rocheteau['charlson_jaccard_score'] = rocheteau_score_list
rocheteau['jaccard_score'] = rocheteau_jaccard_score_list

# Save
#my_method.to_csv('cbr_my_method_metric_charlson_jaccard.csv')
#one_hot.to_csv('cbr_one_hot_metric_charlson_jaccard.csv')
#rocheteau.to_csv('cbr_rocheteau_metric_charlson_jaccard.csv')
